+++ src/oms_sensemaking/config.py
@@ -280,6 +280,15 @@ class ObjectStandardsSettings(BaseModel):
     buffer_expire_sec: int = Field(
         30, description="Delay in seconds to wait before processing Atoms Objects for Object Standards checks")
 
+class COCOMTraversalSettings(BaseModel):
+    """Settings for COCOM Traversal sensemaker"""
+    detect_cocom_traversals: bool = Field(True, description="Toggle on/off COCOM Traversal Detection")
+    rmq_cocom_traversal_queue_name: str = Field(
+        "cocom-traversal-trigger",
+        description="The RMQ COCOM Traversal Queue name",
+        examples=["cocom-traversal-trigger"]
+    )
+
 class IncursionSettings(BaseModel):
     activity_state: str = Field(
         "INCURSION",
@@ -553,6 +562,7 @@ class Settings(BaseSettings):
     object_standards_settings: ObjectStandardsSettings = ObjectStandardsSettings()
     incursion_settings: IncursionSettings = IncursionSettings()
     out_of_garrison_settings: OutOfGarrisonSettings = OutOfGarrisonSettings()
+    cocom_traversal_settings: COCOMTraversalSettings = COCOMTraversalSettings()
     in_port_settings: InPortSettings = InPortSettings()
 
     iw_settings: IWSettings = IWSettings()